requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=5.2.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
blinker=1.7.0=pypi_0
ca-certificates=2023.11.17=h56e8100_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.1=pyhd8ed1ab_0
debugpy=1.6.7=py39hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
filelock=3.13.1=pypi_0
fsspec=2023.12.2=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.41=pypi_0
google-ai-generativelanguage=0.4.0=pypi_0
google-api-core=2.15.0=pypi_0
google-auth=2.26.2=pypi_0
google-generativeai=0.3.2=pypi_0
googleapis-common-protos=1.62.0=pypi_0
grpcio=1.60.0=pypi_0
grpcio-status=1.60.0=pypi_0
huggingface-hub=0.20.2=pypi_0
idna=3.6=pypi_0
importlib-metadata=6.11.0=pypi_0
importlib_metadata=7.0.1=hd8ed1ab_0
ipykernel=6.28.0=pyha63f2e9_0
ipython=8.12.0=pyh08f2357_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
jsonschema=4.21.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=4.12.0=py39hcbf5309_0
libsodium=1.0.18=h8d14728_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
numpy=1.26.3=pypi_0
openssl=3.0.12=h2bbff1b_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.4=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=23.3.1=py39haa95532_0
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
proto-plus=1.23.0=pypi_0
protobuf=4.25.2=pypi_0
psutil=5.9.0=py39h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=14.0.2=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
python=3.9.18=h1aa4202_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2023.3.post1=pypi_0
pywin32=227=py39hb82d6ee_1
pyyaml=6.0.1=pypi_0
pyzmq=25.1.0=py39hd77b12b_0
referencing=0.32.1=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
rich=13.7.0=pypi_0
rpds-py=0.17.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.1=pypi_0
setuptools=68.2.2=py39haa95532_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.28.0=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
tokenizers=0.15.0=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
torch=2.1.2=pypi_0
torchvision=0.16.2=pypi_0
tornado=6.2=py39hb82d6ee_0
tqdm=4.66.1=pypi_0
traitlets=5.14.1=pyhd8ed1ab_0
transformers=4.36.2=pypi_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023.4=pypi_0
tzlocal=5.2=pypi_0
urllib3=2.1.0=pypi_0
validators=0.22.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=3.0.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.41.2=py39haa95532_0
zeromq=4.3.4=h0e60522_1
zipp=3.17.0=pyhd8ed1ab_0
